Refactor: split LegacyReleaseManagement into standalone DuplicateFiles and MissingEpisodes pages
